NIFIREG-267 Moving external controller service references from VersionedProcessGroup to VersionedFlowSnapshot
